feature: split AbstractModel into Model, HasPersistence and HasRelationships concerns
Extracts identity map, fill(), save(), saveRelation() and toDatabase() into the HasPersistence trait, and getRelations(), loadRelation(), loadAllRelations(), relation() and eagerLoad() into the HasRelationships trait. AbstractModel now only keeps construction, magic accessors, casting and array/model conversion, dropping from ~1000 to ~250 lines. Public API is unchanged — AbstractRepository and other consumers continue to work without modification. Duplicate property declarations (withTrashedRelations, onlyTrashedRelations) were removed from AbstractModel since they're now provided by HasRelationships.
