+++ neo/Core/Database/ORM/Model/AbstractModel.php
@@ -3,18 +3,15 @@ declare(strict_types=1);
 
 namespace Neo\Core\Database\ORM\Model;
 
-use Neo\Core\Database\DatabaseConnection;
 use Neo\Core\Database\Exception\DatabaseException;
-use Neo\Core\Database\ORM\Attribute\HasOne;
-use Neo\Core\Database\ORM\Attribute\HasMany;
-use Neo\Core\Database\ORM\Attribute\BelongsTo;
-use Neo\Core\Database\ORM\Attribute\BelongsToMany;
-use Neo\Core\Error\Exception\FrameworkException;
-use PDO;
-use PDOException;
+use Neo\Core\Database\ORM\Model\Concerns\HasPersistence;
+use Neo\Core\Database\ORM\Model\Concerns\HasRelationships;
 
 abstract class AbstractModel
 {
+    use HasPersistence;
+    use HasRelationships;
+
     protected static ?string $table = null;
 
     protected static string $primaryKey = 'id';
@@ -22,27 +19,7 @@ abstract class AbstractModel
     /** @var array<string, mixed> */
     protected array $data = [];
 
-    /** @var array<string, mixed> */
-    protected array $relationsCache = [];
-
-    /** @var array<string, AbstractModel> */
-    private static array $instanceCache = [];
-
-    protected bool $withTrashedRelations = false;
-
-    protected bool $onlyTrashedRelations = false;
-
-    /** @var array<int, string> */
-    protected array $fillable = [];
-
-    protected bool $trackIdentity = true;
-
-    protected bool $timestamps = true;
-
     protected bool $isLoadingRelations = false;
-
-    /** @var array<int, string> */
-    protected static array $loadingGuard = [];
 
     /** @var array<int, string> */
     protected array $hidden = [];
@@ -89,19 +66,6 @@ abstract class AbstractModel
     }
 
     /**
-     * @return array<string, mixed>
-     */
-    public function getRelationsCache(): array
-    {
-        return $this->relationsCache;
-    }
-
-    public static function clearIdentityMap(): void
-    {
-        self::$instanceCache = [];
-    }
-
-    /**
      * @return array<int, string>
      */
     private function getInternalProperties(): array
@@ -117,38 +81,6 @@ abstract class AbstractModel
             'isLoadingRelations',
             'hidden',
         ];
-    }
-
-    public function registerIdentity(): void
-    {
-        if (!$this->trackIdentity) return;
-
-        $pk = static::getPrimaryKey();
-        $id = $this->$pk ?? null;
-
-        if ($id === null || $id === 0 || $id === '0' || $id === '') return;
-
-        $key = static::class . ':' . $id;
-        self::$instanceCache[$key] = $this;
-    }
-
-    /**
-     * @return static|null
-     */
-    public static function getIdentity(int|string $id): ?self
-    {
-        $key = static::class . ':' . $id;
-        if (!isset(self::$instanceCache[$key])) {
-            return null;
-        }
-
-        /** @var static */
-        return self::$instanceCache[$key];
-    }
-
-    public static function removeIdentity(int|string $id): void
-    {
-        unset(self::$instanceCache[static::class . ':' . $id]);
     }
 
     /**
@@ -183,11 +115,6 @@ abstract class AbstractModel
         $this->data[$name] = $value;
     }
 
-    public function usesTimestamps(): bool
-    {
-        return $this->timestamps;
-    }
-
     public static function getTable(): string
     {
         return static::$table
@@ -197,280 +124,6 @@ abstract class AbstractModel
     public static function getPrimaryKey(): string
     {
         return static::$primaryKey;
-    }
-
-    /**
-     * @return array<string, object>
-     */
-    public function getRelations(): array
-    {
-        static $cache = [];
-        $class = static::class;
-
-        if (isset($cache[$class])) return $cache[$class];
-
-        $ref = new \ReflectionClass($class);
-        $relations = [];
-
-        foreach ($ref->getProperties() as $prop) {
-
-            $attrs = [
-                ...$prop->getAttributes(HasOne::class),
-                ...$prop->getAttributes(HasMany::class),
-                ...$prop->getAttributes(BelongsTo::class),
-                ...$prop->getAttributes(BelongsToMany::class),
-            ];
-
-            foreach ($attrs as $attr) {
-                $relations[$prop->getName()] = $attr->newInstance();
-            }
-        }
-
-        return $cache[$class] = $relations;
-    }
-
-    protected function hasRelation(string $name): bool
-    {
-        return isset($this->getRelations()[$name]);
-    }
-
-    /**
-     * @param array<int, AbstractModel> $items
-     * @return array<int, AbstractModel>
-     */
-    protected function filterTrashed(array $items, bool $include, bool $only): array
-    {
-        return array_values(array_filter($items, function ($item) use ($include, $only) {
-            if (!$item instanceof self) return true;
-            /** @phpstan-ignore-next-line */
-            if ($only) return $item->deleted_at !== null;
-            /** @phpstan-ignore-next-line */
-            if (!$include) return $item->deleted_at === null;
-            return true;
-        }));
-    }
-
-    /**
-     * @throws \ReflectionException
-     */
-    public static function getRelationTarget(string $relationName): ?string
-    {
-        $ref  = new \ReflectionClass(static::class);
-        $prop = $ref->getProperty($relationName);
-
-        $attrs = [
-            ...$prop->getAttributes(HasOne::class),
-            ...$prop->getAttributes(HasMany::class),
-            ...$prop->getAttributes(BelongsTo::class),
-            ...$prop->getAttributes(BelongsToMany::class),
-        ];
-
-        return !empty($attrs) ? $attrs[0]->newInstance()->target : null;
-    }
-
-    /**
-     * @throws DatabaseException
-     */
-    public function loadRelation(
-        string $name,
-        bool $includeTrashed = false,
-        bool $onlyTrashed = false,
-        bool $forceReload = false
-    ): mixed {
-        if (!$forceReload && array_key_exists($name, $this->relationsCache)) {
-            return $this->relationsCache[$name];
-        }
-
-        $pk = static::getPrimaryKey();
-        $guardKey = static::class . ':' . ($this->$pk ?? spl_object_id($this)) . ':' . $name;
-
-        if (!$forceReload && isset(self::$loadingGuard[$guardKey])) {
-            return null;
-        }
-
-        self::$loadingGuard[$guardKey] = true;
-
-        try {
-            $relations = $this->getRelations();
-            if (!isset($relations[$name])) {
-                $this->relationsCache[$name] = null;
-                return null;
-            }
-
-            $relation = $relations[$name];
-            $pdo = DatabaseConnection::getPdo();
-            $where = $onlyTrashed
-                ? 'deleted_at IS NOT NULL'
-                : ($includeTrashed ? '1=1' : 'deleted_at IS NULL');
-
-            try {
-                if ($relation instanceof BelongsTo) {
-                    $fk = $this->{$relation->foreignKey};
-                    if (!$fk) {
-                        $this->relationsCache[$name] = null;
-                        return null;
-                    }
-
-                    $stmt = $pdo->prepare("SELECT * FROM {$relation->target::getTable()} WHERE {$relation->target::getPrimaryKey()} = ? AND $where LIMIT 1");
-                    $stmt->execute([$fk]);
-                    $row = $stmt->fetch(PDO::FETCH_ASSOC);
-                    $result = $row ? $relation->target::hydrateRow($row) : null;
-
-                    $this->relationsCache[$name] = $result;
-                    return $result;
-                }
-
-                if ($relation instanceof HasOne) {
-                    $stmt = $pdo->prepare("SELECT * FROM {$relation->target::getTable()} WHERE {$relation->foreignKey} = ? AND $where LIMIT 1");
-                    $stmt->execute([$this->{$relation->localKey}]);
-                    $row = $stmt->fetch(PDO::FETCH_ASSOC);
-                    $result = $row ? $relation->target::hydrateRow($row) : null;
-
-                    $this->relationsCache[$name] = $result;
-                    return $result;
-                }
-
-                if ($relation instanceof HasMany) {
-                    $stmt = $pdo->prepare("SELECT * FROM {$relation->target::getTable()} WHERE {$relation->foreignKey} = ? AND $where");
-                    $stmt->execute([$this->{$relation->localKey}]);
-                    $rows = $stmt->fetchAll(PDO::FETCH_ASSOC) ?: [];
-                    $objs = array_map(fn($r) => $relation->target::hydrateRow($r), $rows);
-
-                    $this->relationsCache[$name] = $objs;
-                    return $objs;
-                }
-
-                if ($relation instanceof BelongsToMany) {
-                    $sql = "
-                        SELECT t.* FROM {$relation->target::getTable()} t
-                        INNER JOIN {$relation->pivotTable} p
-                            ON p.{$relation->pivotTargetKey} = t.{$relation->target::getPrimaryKey()}
-                        WHERE p.{$relation->pivotLocalKey} = ?
-                          AND " . ($onlyTrashed
-                            ? 't.deleted_at IS NOT NULL'
-                            : ($includeTrashed ? '1=1' : 't.deleted_at IS NULL'));
-
-                    $stmt = $pdo->prepare($sql);
-                    $stmt->execute([$this->{$relation->localKey}]);
-                    $rows = $stmt->fetchAll(PDO::FETCH_ASSOC) ?: [];
-                    $objs = array_map(fn($r) => $relation->target::hydrateRow($r), $rows);
-
-                    $this->relationsCache[$name] = $objs;
-                    return $objs;
-                }
-            } catch (PDOException $e) {
-                throw new DatabaseException(
-                    title: 'Relation Load Error',
-                    message: sprintf("Error while loading relation '%s' on %s: %s", $name, static::class, $e->getMessage()),
-                    code: 500,
-                    previous: $e
-                );
-            }
-
-            $this->relationsCache[$name] = null;
-            return null;
-
-        } finally {
-            unset(self::$loadingGuard[$guardKey]);
-        }
-    }
-
-    /**
-     * @param array<string, mixed> $row
-     * @throws DatabaseException
-     */
-    public static function hydrateRow(array $row): static
-    {
-        return new static($row);
-    }
-
-    /**
-     * @param array<int, string> $relations
-     * @throws DatabaseException
-     */
-    public function loadRelations(array $relations): void
-    {
-        foreach ($relations as $name) {
-            if ($this->hasRelation($name)) {
-                $this->relationsCache[$name] = $this->loadRelation(
-                    $name,
-                    $this->withTrashedRelations,
-                    $this->onlyTrashedRelations
-                );
-            }
-        }
-    }
-
-    /**
-     * @param array<string, bool> $loaded
-     * @throws DatabaseException
-     */
-    public function loadAllRelations(
-        bool $recursive = true,
-        array &$loaded = [],
-        int $depth = 0,
-        int $maxDepth = 5,
-    ): void {
-        $pk = static::getPrimaryKey();
-        $key = static::class . ':' . ($this->$pk ?? spl_object_id($this));
-
-        if (isset($loaded[$key]) || $depth >= $maxDepth) return;
-
-        $loaded[$key] = true;
-
-        foreach ($this->getRelations() as $name => $_) {
-            $this->relationsCache[$name] = $this->loadRelation(
-                $name,
-                $this->withTrashedRelations,
-                $this->onlyTrashedRelations
-            );
-
-            if (!$recursive) continue;
-
-            $rel = $this->relationsCache[$name];
-
-            if ($rel instanceof self) {
-                $rel->loadAllRelations(true, $loaded, $depth + 1, $maxDepth);
-            } elseif (is_array($rel)) {
-                foreach ($rel as $obj) {
-                    if ($obj instanceof self) {
-                        $obj->loadAllRelations(true, $loaded, $depth + 1, $maxDepth);
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * @throws DatabaseException
-     */
-    public function relation(string $name): mixed
-    {
-        if (!$this->hasRelation($name)) return null;
-
-        if (!array_key_exists($name, $this->relationsCache)) {
-            $this->relationsCache[$name] = $this->loadRelation(
-                $name,
-                $this->withTrashedRelations,
-                $this->onlyTrashedRelations
-            );
-        }
-
-        return $this->relationsCache[$name];
-    }
-
-    public function withTrashedRelations(): self
-    {
-        $this->withTrashedRelations = true;
-        $this->onlyTrashedRelations = false;
-        return $this;
-    }
-
-    public function onlyTrashedRelations(): self
-    {
-        $this->withTrashedRelations = true;
-        $this->onlyTrashedRelations = true;
-        return $this;
     }
 
     /**
@@ -548,267 +201,6 @@ abstract class AbstractModel
         return $this;
     }
 
-    /**
-     * @param array<string, mixed> $data
-     */
-    public function fill(array $data): void
-    {
-        $relations = array_keys($this->getRelations());
-
-        foreach ($data as $key => $value) {
-            if (!empty($this->fillable) && !in_array($key, $this->fillable, true)) continue;
-            if (in_array($key, $relations, true)) continue;
-
-            if (property_exists($this, $key)) {
-                $prop = new \ReflectionProperty($this, $key);
-                $prop->setValue($this, $value);
-            }
-
-            $this->data[$key] = $value;
-        }
-    }
-
-    public function save(): bool
-    {
-        try {
-            $pdo = DatabaseConnection::getPdo();
-            $data = $this->toDatabase();
-            $pk = static::getPrimaryKey();
-            $table = static::getTable();
-
-            foreach ($this->getRelations() as $relName => $_) {
-                unset($data[$relName]);
-            }
-
-            static $columnsCache = [];
-            if (!isset($columnsCache[$table])) {
-                $stmt = $pdo->query("DESCRIBE `$table`");
-                if ($stmt === false) {
-                    throw new DatabaseException(
-                        title: 'Model Save Error',
-                        message: sprintf("Unable to describe table '%s'.", $table),
-                        code: 500
-                    );
-                }
-                $columnsCache[$table] = $stmt->fetchAll(PDO::FETCH_COLUMN);
-            }
-
-            $tableColumns = $columnsCache[$table];
-            $isInsert = empty($this->$pk);
-            $now = (new \DateTime())->format('Y-m-d H:i:s');
-
-            if ($isInsert) {
-                if (in_array('created_at', $tableColumns) && !isset($data['created_at'])) {
-                    $data['created_at'] = $now;
-                }
-                if (in_array('updated_at', $tableColumns) && !isset($data['updated_at'])) {
-                    $data['updated_at'] = $now;
-                }
-            } else {
-                if (in_array('updated_at', $tableColumns)) {
-                    $data['updated_at'] = $now;
-                }
-            }
-
-            if ($isInsert) {
-                unset($data[$pk]);
-                $columns = implode(', ', array_map(fn($c) => "`$c`", array_keys($data)));
-                $placeholders = implode(', ', array_fill(0, count($data), '?'));
-                $stmt = $pdo->prepare("INSERT INTO `$table` ($columns) VALUES ($placeholders)");
-                $stmt->execute(array_values($data));
-                $this->$pk = (int) $pdo->lastInsertId();
-            } else {
-                $set = implode(', ', array_map(fn($col) => "`$col` = ?", array_keys($data)));
-                $stmt = $pdo->prepare("UPDATE `$table` SET $set WHERE `$pk` = ?");
-                $stmt->execute([...array_values($data), $this->$pk]);
-            }
-
-            $this->registerIdentity();
-            return true;
-
-        } catch (FrameworkException $e) {
-            throw $e;
-        } catch (\Throwable $e) {
-            throw new DatabaseException(
-                title: 'Model Save Error',
-                message: sprintf("Error while saving model '%s': %s", static::class, $e->getMessage()),
-                code: 500,
-                previous: $e
-            );
-        }
-    }
-
-    /**
-     * @param array<int, AbstractModel|array<string, mixed>> $entries
-     * @throws DatabaseException
-     */
-    public function saveRelation(string $relationName, array $entries): void
-    {
-        $relations = $this->getRelations();
-
-        if (!isset($relations[$relationName])) {
-            throw new DatabaseException(
-                title: 'Model Relation Error',
-                message: sprintf("Relation '%s' does not exist on %s.", $relationName, static::class),
-                code: 500
-            );
-        }
-
-        $relation = $relations[$relationName];
-
-        if (!$relation instanceof HasMany) {
-            throw new DatabaseException(
-                title: 'Model Relation Error',
-                message: sprintf("saveRelation() only supports HasMany relations. '%s' is not HasMany.", $relationName),
-                code: 500
-            );
-        }
-
-        $pdo = DatabaseConnection::getPdo();
-        $target = $relation->target;
-        $foreignKey = $relation->foreignKey;
-        $localKey = $relation->localKey;
-        $targetPk = $target::getPrimaryKey();
-        $table = $target::getTable();
-        $parentId = $this->$localKey;
-
-        $submittedIds = [];
-        foreach ($entries as $entry) {
-            $pk = is_object($entry)
-                ? $entry->$targetPk
-                : ($entry[$targetPk] ?? null);
-
-            if ($pk) $submittedIds[] = (int)$pk;
-        }
-
-        $stmt = $pdo->prepare(
-            "SELECT `{$targetPk}` FROM `{$table}` WHERE `{$foreignKey}` = ? AND deleted_at IS NULL"
-        );
-
-        $stmt->execute([$parentId]);
-        $existingIds = array_column($stmt->fetchAll(\PDO::FETCH_ASSOC), $targetPk);
-        $existingIds = array_map('intval', $existingIds);
-
-        $toDelete = array_diff($existingIds, $submittedIds);
-
-        if (!empty($toDelete)) {
-            $placeholders = implode(', ', array_fill(0, count($toDelete), '?'));
-            $now = date('Y-m-d H:i:s');
-
-            $hasSoftDelete = $this->targetHasSoftDelete($target, $pdo);
-
-            if ($hasSoftDelete) {
-                $stmt = $pdo->prepare(
-                    "UPDATE `{$table}` SET deleted_at = ? WHERE `{$targetPk}` IN ({$placeholders})"
-                );
-                $stmt->execute([$now, ...$toDelete]);
-            } else {
-                $stmt = $pdo->prepare(
-                    "DELETE FROM `{$table}` WHERE `{$targetPk}` IN ({$placeholders})"
-                );
-                $stmt->execute([...$toDelete]);
-            }
-
-            foreach ($toDelete as $deletedId) {
-                $target::removeIdentity($deletedId);
-            }
-        }
-
-        $now = date('Y-m-d H:i:s');
-
-        foreach ($entries as $entry) {
-            if (!$entry instanceof AbstractModel) {
-                $entry = new $target(is_array($entry) ? $entry : []);
-            }
-
-            $entry->$foreignKey = $parentId;
-
-            $pk = $entry->$targetPk ?? null;
-
-            if ($pk) {
-                $data = $entry->toDatabase();
-                unset(
-                    $data[$targetPk],
-                    $data['created_at'],
-                    $data['deleted_at']
-                );
-
-                $data['updated_at'] = $now;
-
-                $setParts = array_map(fn($col) => "`{$col}` = ?", array_keys($data));
-                $stmt = $pdo->prepare(
-                    "UPDATE `{$table}` SET " . implode(', ', $setParts) . " WHERE `{$targetPk}` = ?"
-                );
-                $stmt->execute([...array_values($data), $pk]);
-
-                $entry->registerIdentity();
-            } else {
-                $data = $entry->toDatabase();
-                unset($data[$targetPk]);
-
-                $data['created_at'] ??= $now;
-                $data['updated_at'] ??= $now;
-
-                $columns = implode(', ', array_map(fn($c) => "`$c`", array_keys($data)));
-                $placeholders = implode(', ', array_fill(0, count($data), '?'));
-
-                $stmt = $pdo->prepare(
-                    "INSERT INTO `{$table}` ({$columns}) VALUES ({$placeholders})"
-                );
-                $stmt->execute(array_values($data));
-
-                $entry->$targetPk = (int)$pdo->lastInsertId();
-                $entry->registerIdentity();
-            }
-        }
-
-        unset($this->relationsCache[$relationName]);
-    }
-
-    private function targetHasSoftDelete(string $target, \PDO $pdo): bool
-    {
-        static $cache = [];
-
-        if (isset($cache[$target])) {
-            return $cache[$target];
-        }
-
-        $table = $target::getTable();
-        $stmt = $pdo->prepare("SHOW COLUMNS FROM {$table} LIKE 'deleted_at'");
-        $stmt->execute();
-
-        return $cache[$target] = (bool)$stmt->fetch();
-    }
-
-    /**
-     * @return array<string, mixed>
-     */
-    public function toDatabase(): array
-    {
-        $data = [];
-        $ref = new \ReflectionObject($this);
-        $relations = array_keys($this->getRelations());
-
-        foreach ($ref->getProperties() as $prop) {
-            if ($prop->isStatic()) continue;
-
-            $name = $prop->getName();
-            if (in_array($name, $this->getInternalProperties(), true)) continue;
-            if (in_array($name, $relations, true)) continue;
-
-            $value = $prop->getValue($this);
-
-            if ($value instanceof self) continue;
-            if (is_array($value) && !empty($value) && $value[0] instanceof self) continue;
-
-            $data[$name] = $value instanceof \DateTime
-                ? $value->format('Y-m-d H:i:s')
-                : $value;
-        }
-
-        return $data;
-    }
-
     protected function castValue(\ReflectionProperty $prop, mixed $value): mixed
     {
         $type = $prop->getType();
@@ -855,146 +247,5 @@ abstract class AbstractModel
         return $onlyTrashed
             ? "$column IS NOT NULL"
             : ($includeTrashed ? '1=1' : "$column IS NULL");
-    }
-
-    /**
-     * @param array<int, AbstractModel> $models
-     * @throws DatabaseException
-     */
-    public static function eagerLoad(
-        array $models,
-        string $relation,
-        bool $includeTrashed = false,
-        bool $onlyTrashed = false
-    ): void {
-        if (empty($models)) return;
-
-        $first = $models[0];
-        $relations = $first->getRelations();
-
-        if (!isset($relations[$relation])) return;
-
-        $rel = $relations[$relation];
-        $pdo = DatabaseConnection::getPdo();
-        $where = self::buildSoftDeleteWhere($includeTrashed, $onlyTrashed);
-
-        $syncWithIdentityMap = function (self $obj): self {
-            return $obj;
-        };
-
-        try {
-            if ($rel instanceof HasMany || $rel instanceof HasOne) {
-                $localKey = $rel->localKey;
-                $foreignKey = $rel->foreignKey;
-                $target = $rel->target;
-
-                $ids = array_values(array_unique(array_filter(
-                    array_map(fn($m) => $m->$localKey, $models)
-                )));
-
-                if (!$ids) return;
-
-                $placeholders = implode(',', array_fill(0, count($ids), '?'));
-                $stmt = $pdo->prepare("SELECT * FROM {$target::getTable()} WHERE $foreignKey IN ($placeholders) AND $where");
-                $stmt->execute($ids);
-                $rows = $stmt->fetchAll(PDO::FETCH_ASSOC);
-                $grouped = [];
-
-                foreach ($rows as $row) {
-                    $obj = $syncWithIdentityMap($target::hydrateRow($row));
-                    $grouped[$row[$foreignKey]][] = $obj;
-                }
-
-                foreach ($models as $model) {
-                    $key = $model->$localKey;
-                    $model->relationsCache[$relation] = $rel instanceof HasOne
-                        ? ($grouped[$key][0] ?? null)
-                        : ($grouped[$key] ?? []);
-                }
-
-                return;
-            }
-
-            if ($rel instanceof BelongsTo) {
-                $foreignKey = $rel->foreignKey;
-                $target = $rel->target;
-                $targetPk = $target::getPrimaryKey();
-
-                $ids = array_values(array_unique(array_filter(
-                    array_map(fn($m) => $m->$foreignKey, $models)
-                )));
-
-                if (!$ids) return;
-
-                $placeholders = implode(',', array_fill(0, count($ids), '?'));
-                $stmt = $pdo->prepare("SELECT * FROM {$target::getTable()} WHERE $targetPk IN ($placeholders) AND $where");
-                $stmt->execute($ids);
-                $rows = $stmt->fetchAll(PDO::FETCH_ASSOC);
-                $mapped = [];
-
-                foreach ($rows as $row) {
-                    $obj = $syncWithIdentityMap($target::hydrateRow($row));
-                    $mapped[$row[$targetPk]] = $obj;
-                }
-
-                foreach ($models as $model) {
-                    $model->relationsCache[$relation] = $mapped[$model->$foreignKey] ?? null;
-                }
-
-                return;
-            }
-
-            if ($rel instanceof BelongsToMany) {
-                $localKey = $rel->localKey;
-                $target = $rel->target;
-                $pivotTable = $rel->pivotTable;
-                $pivotLocalKey = $rel->pivotLocalKey;
-                $pivotTargetKey = $rel->pivotTargetKey;
-                $targetPk = $target::getPrimaryKey();
-
-                $ids = array_values(array_unique(array_filter(
-                    array_map(fn($m) => $m->$localKey, $models)
-                )));
-
-                if (!$ids) return;
-
-                $placeholders = implode(',', array_fill(0, count($ids), '?'));
-                $sql = "
-                    SELECT t.*, p.$pivotLocalKey AS pivot_local
-                    FROM {$target::getTable()} t
-                    INNER JOIN $pivotTable p ON p.$pivotTargetKey = t.$targetPk
-                    WHERE p.$pivotLocalKey IN ($placeholders)
-                    AND $where
-                ";
-
-                $stmt = $pdo->prepare($sql);
-                $stmt->execute($ids);
-                $rows = $stmt->fetchAll(PDO::FETCH_ASSOC);
-                $grouped = [];
-
-                foreach ($rows as $row) {
-                    $obj = $syncWithIdentityMap($target::hydrateRow($row));
-                    $grouped[$row['pivot_local']][] = $obj;
-                }
-
-                foreach ($models as $model) {
-                    $model->relationsCache[$relation] = $grouped[$model->$localKey] ?? [];
-                }
-
-                return;
-            }
-
-        } catch (PDOException $e) {
-            throw new DatabaseException(
-                title: 'Eager Load Error',
-                message: sprintf("Error while eager loading relation '%s': %s", $relation, $e->getMessage()),
-                code: 500,
-                previous: $e
-            );
-        }
-
-        foreach ($models as $model) {
-            $model->loadRelation($relation, $includeTrashed, $onlyTrashed);
-        }
     }
 }
